manualAppProject/ny_db_SQLA/nginx_o_files_from_ec2/requirements.txt
# Web Framework & API
fastapi
uvicorn
pydantic
pydantic-settings
email-validator

# Database
sqlalchemy
alembic
psycopg2

# AWS
boto3
botocore
awscli

# Authentication & Security
python-jose
passlib
bcrypt
python-multipart

# Document Processing
PyMuPDF
pymupdf4llm
easyocr
opencv-python
numpy
requests

# Text Processing & Detection
langdetect

# AI Services
mistralai

# Utilities
python-dotenv